+++ tests/tensor/test_split.py
@@ -13,7 +13,7 @@ def svd_combine(a):
     US = yast.tensordot(U, S, axes=(2, 0))
     USV = yast.tensordot(US, V, axes=(2, 0))
     USV = USV.transpose(axes=(3, 1, 2, 0))
-    assert a.norm_diff(USV) < tol  # == 0.0
+    assert yast.norm(a - USV) < tol  # == 0.0
     assert a.is_consistent()
     assert U.is_consistent()
     assert S.is_consistent()
@@ -25,7 +25,7 @@ def svd_order_combine(a):
     US = yast.tensordot(S, U, axes=(0, 0))
     USV = yast.tensordot(US, V, axes=(0, 2))
     USV = USV.transpose(axes=(3, 1, 2, 0))
-    assert a.norm_diff(USV) < tol  # == 0.0
+    assert yast.norm(a - USV) < tol  # == 0.0
     assert U.is_consistent()
     assert S.is_consistent()
     assert V.is_consistent()
@@ -35,7 +35,7 @@ def qr_combine(a):
     Q, R = yast.linalg.qr(a, axes=((3, 1), (2, 0)))
     QR = yast.tensordot(Q, R, axes=(2, 0))
     QR = QR.transpose(axes=(3, 1, 2, 0))
-    assert a.norm_diff(QR) < tol  # == 0.0
+    assert yast.norm(a - QR) < tol  # == 0.0
     assert Q.is_consistent()
     assert R.is_consistent()
 
@@ -44,7 +44,7 @@ def qr_order_combine(a):
     Q, R = yast.linalg.qr(a, axes=((3, 1), (2, 0)), sQ=-1, Qaxis=-2, Raxis=1)
     QR = yast.tensordot(Q, R, axes=(1, 1))
     QR = QR.transpose(axes=(3, 1, 2, 0))
-    assert a.norm_diff(QR) < tol  # == 0.0
+    assert yast.norm(a - QR) < tol  # == 0.0
     assert Q.is_consistent()
     assert R.is_consistent()
 
@@ -54,7 +54,7 @@ def eigh_combine(a):
     S, U = yast.linalg.eigh(a2, axes=((0, 1), (2, 3)))
     US = yast.tensordot(U, S, axes=(2, 0))
     USU = yast.tensordot(US, U, axes=(2, 2), conj=(0, 1))
-    assert a2.norm_diff(USU) < tol  # == 0.0
+    assert yast.norm(a2 - USU) < tol  # == 0.0
     assert U.is_consistent()
     assert S.is_consistent()
 
@@ -64,7 +64,7 @@ def eigh_order_combine(a):
     S, U = yast.linalg.eigh(a2, axes=((0, 1), (2, 3)), Uaxis=0, sU=-1)
     US = yast.tensordot(S, U, axes=(0, 0))
     USU = yast.tensordot(US, U, axes=(0, 0), conj=(0, 1))
-    assert a2.norm_diff(USU) < tol  # == 0.0
+    assert yast.norm(a2 - USU) < tol  # == 0.0
     assert U.is_consistent()
     assert S.is_consistent()